refactor(view): create and use birth-specific subclass of `DateView`
