+++ web/vital_records/views/common.py
@@ -152,21 +152,11 @@ class CountyView(StepsMixin, EligibilityMixin, ValidateRequestIdMixin, UpdateVie
         return super().form_valid(form)
 
 
-class DateOfBirthView(StepsMixin, EligibilityMixin, ValidateRequestIdMixin, UpdateView):
+class DateOfEventView(StepsMixin, EligibilityMixin, ValidateRequestIdMixin, UpdateView):
     model = VitalRecordsRequest
-    form_class = DateOfBirthForm
     template_name = "vital_records/request/form.html"
     context_object_name = "vital_request"
     step_name = Steps.date_of_birth
-
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
-        context["form_layout"] = "date_form"
-        context["font_hint_name"] = "dob-hint"
-        context["form_question"] = "What is the date of birth?"
-        context["form_hint"] = "If you’re not sure, enter your approximate date of birth."
-
-        return context
 
     def form_valid(self, form):
         # Move form state to next state
